Separated ModCheckboxTable from GameOptionsTable, so we can use it when creating a map
